Extract marc utilities to common module
